Renamed ChessServer\GameMode as ChessServer\Game
