+++ src/Game/GameModeStorage.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace ChessServer\GameMode;
+namespace ChessServer\Game;
 
 class GameModeStorage extends \SplObjectStorage
 {
