+++ src/Game/SanMode.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace ChessServer\GameMode;
+namespace ChessServer\Game;
 
 use ChessServer\Game;
